requirements.txt
# Flask and extensions
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
requests==2.28.2

# Google API
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.81.0

# NLP libraries (lightweight)
spacy==3.5.1
scikit-learn==1.2.2

# Download lightweight spaCy model:
# python -m spacy download en_core_web_sm

# Utils
python-dateutil==2.8.2
pytz==2023.3